VFB-171: Rename getCurrentYear file
