+++ src/app/clients/form/ClientForm.tsx
@@ -36,7 +36,7 @@ import SignpostingCallCard from "@/app/clients/form/formSections/SignpostingCall
 import Button from "@mui/material/Button";
 import { submitAddClientForm, submitEditClientForm } from "@/app/clients/form/submitFormHelpers";
 import Title from "@/components/Title/Title";
-import { getCurrentYear } from "@/common/getCurrentYear";
+import { getCurrentYear } from "@/common/getAgesOfFamily";
 import { adultBirthYearList } from "@/app/clients/form/birthYearDropdown";
 
 interface Props {
